chore: 디렉터리 구조 변경
초대링크인 post와 사용자가 내역을 조회할 수 있는 letter를 구분하기위함
